requirements.txt
discord.py
cleverbotfree
python-dotenv
